+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Added
 * Added new `MonologFactory` for monolog 3, and a couple new logger utilities.
 * Added new `RabbitMqPublishingHelper`, to simplify publishing in rabbitmq queues.
+* Added new `RedisPublishingHelper`, to simplify publishing in redis pub/sub queues.
 
 ### Changed
 * Upgraded to predis 2.0.0